interviewList 평균값 null로받을경우 수정

+++ src/main/java/com/signal/interview/dao/InterviewDAO.java
@@ -22,7 +22,7 @@ public interface InterviewDAO {
 
 	ArrayList<InterviewDTO> interviewDetailResultList(String inter_no);
 
-	float avgGrade(String cl_id);
+	int avgGrade(String cl_id);
 
 	int countComment(String cl_id);
